+++ Realm/RLMSyncSubscription_Private.h
@@ -56,6 +56,9 @@ NS_ASSUME_NONNULL_BEGIN
                            predicate:(NSPredicate *)predicate
                       updateExisting:(BOOL)updateExisting;
 
+- (void)waitForSynchronizationOnQueue:(nullable dispatch_queue_t)queue
+                      completionBlock:(void(^)(NSError *))completionBlock;
+
 - (RLMSyncSubscriptionEnumerator *)fastEnumerator;
 
 NSUInteger RLMFastEnumerate(NSFastEnumerationState *state,